@@ -4,10 +4,10 @@ class Translators::YandexWrapper < Translators::Wrapper
   TRANSLATOR = Yandex::Translator.new(Rails.application.secrets.yandex_translator)
 
   def translate(from, to)
-    @translation ||= TRANSLATOR.translate(@text, from: from, to: to)
+    TRANSLATOR.translate(@text, from: from, to: to)
   end
 
   def detect
-    @detected ||= TRANSLATOR.detect(@text)
+    @detect ||= TRANSLATOR.detect(@text)
   end
 end
