@@ -8,7 +8,7 @@ require 'telegram/bot'
 require 'telegram/bot/railtie'
 require 'telegram/bot/types'
 
-require File.expand_path('../../config/environment', __FILE__)
+require File.expand_path('../config/environment', __dir__)
 
 abort('The Rails environment is running in production mode!') if Rails.env.production?
 require 'rspec/rails'
